More efficient broad phase of collision detection.
src/main.cpp: Now using a hash table instead of an array to keep track of pairs of actors that have already been checked in the current frame. The keys are the XORed pointers of the current two meshes, and the values are boolean and either true if the pair of actors whose XORed pointers correspond to the keys have been checked before, or false otherwise.